pyproject.toml
[tool.poetry]
name = "finance-gpt"
version = "0.2.0"
description = ""
authors = ["Danil Ustinov <vk.vx@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.1.7"
langchain-community = "^0.0.20"
langchain-core = "^0.1.23"
langchain-openai = "^0.0.6"
brotli = "^1.1.0"
python-telegram-bot = "^20.8"
gspread = "^6.0.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
mypy = "^1.7.0"
make-better = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
